moved sbt plugin files
