+++ stryker4jvm-plugin-sbt/src/main/scala/stryker4jvm/plugin/sbt/Stryker4jvmMain.scala
@@ -1,4 +1,4 @@
-package stryker4s.sbt
+package stryker4jvm.plugin.sbt
 
 import cats.effect.unsafe.IORuntime
 import cats.effect.{Deferred, IO}
@@ -6,14 +6,15 @@ import sbt.Keys.*
 import sbt.*
 import sbt.plugins.*
 import stryker4jvm.core.logging.Logger
-import stryker4s.log.SbtLogger
 import stryker4jvm.run.threshold.ErrorStatus
+
 import scala.concurrent.duration.FiniteDuration
 import fs2.io.file
+import stryker4jvm.plugin.sbt.logging.FansiSbtLogger
 
 /** This plugin adds a new task (stryker) to the project that allow you to run mutation testing over your code
   */
-object Stryker4sMain extends AutoPlugin {
+object Stryker4jvmMain extends AutoPlugin {
   override def requires = JvmPlugin
 
   override def trigger = allRequirements
@@ -43,14 +44,14 @@ object Stryker4sMain extends AutoPlugin {
     val _ = (stryker / logLevel).value
 
     implicit val runtime: IORuntime = IORuntime.global
-    implicit val logger: Logger = new SbtLogger(streams.value.log)
+    implicit val logger: Logger = new FansiSbtLogger(streams.value.log).logger
 
     val sources =
       Seq((Compile / scalaSource).value, (Compile / javaSource).value).map(_.toPath()).map(file.Path.fromNioPath)
     val targetPath = file.Path.fromNioPath(target.value.toPath())
 
     Deferred[IO, FiniteDuration] // Create shared timeout between testrunners
-      .map(new Stryker4sSbtRunner(state.value, _, sources, targetPath))
+      .map(new Stryker4jvmSbtRunner(state.value, _, sources, targetPath))
       .flatMap(_.run())
       .map {
         case ErrorStatus => throw new MessageOnlyException("Mutation score is below configured threshold")
