+++ stryker4jvm-plugin-sbt/src/main/scala/stryker4jvm/plugin/sbt/files/SbtFilesResolver.scala
@@ -1,10 +1,10 @@
-package stryker4s.files
+package stryker4jvm.plugin.sbt.files
 
 import cats.effect.IO
 import fs2.Stream
 import fs2.io.file.{Files, Path}
-import stryker4jvm.files.FilesFileResolver
 import stryker4jvm.extensions.StreamExtensions.*
+import stryker4jvm.files.FilesFileResolver
 
 class SbtFilesResolver(sources: Seq[Path], target: Path) extends FilesFileResolver {
