+++ stryker4jvm-plugin-sbt/src/main/scala/stryker4jvm/plugin/sbt/runner/TestRunnerConnection.scala
@@ -1,4 +1,4 @@
-package stryker4s.sbt.runner
+package stryker4jvm.plugin.sbt.runner
 
 import cats.effect.IO
 import cats.syntax.bifunctor.*
